third_party/blink/web_tests/external/wpt/css/fill-stroke/text-decoration-stroke/text-decoration-stroke-computed.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-decoration-stroke value 'none'
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke value 'context-fill'
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke value 'context-stroke'
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke value 'rgb(12, 34, 56)'
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke value 'url("https://example.com/")'
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke value 'url("https://example.com/") none'
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-stroke value 'url("https://example.com/") rgb(12, 34, 56)'
  assert_true: text-decoration-stroke doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
